pagination cluster status fix
